yolov8/requirements.txt
ultralytics==8.3.141
flask==3.0.0
werkzeug==3.0.1
opencv-python==4.9.0.80
pillow==10.4.0
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0
requests==2.32.3